deleted blue buttons and made it so that logout renders back to welcome page
