+++ Tabloid/client/src/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { BrowserRouter as Router } from "react-router-dom";
 import { UserProfileProvider } from "./providers/UserProfileProvider";
-import Header from "./components/views/Header";
-import ApplicationViews from "./components/views/ApplicationViews";
+import Header from "./components/Header";
+import ApplicationViews from "./components/ApplicationViews";
 import { CategoryProvider } from './providers/CategoryProvider';
 import { PostProvider } from './providers/PostProvider';
