+++ Tabloid/client/src/components/Register.js
@@ -1,12 +1,11 @@
 import React, { useState, useContext } from "react";
 import { Button, Form, FormGroup, Label, Input } from 'reactstrap';
 import { useHistory } from "react-router-dom";
-import { UserProfileContext } from "../../providers/UserProfileProvider";
+import { UserProfileContext } from "../providers/UserProfileProvider";
 
 export default function Register() {
   const history = useHistory();
   const { register } = useContext(UserProfileContext);
-
   const [firstName, setFirstName] = useState();
   const [lastName, setLastName] = useState();
   const [displayName, setDisplayName] = useState();
@@ -24,7 +23,7 @@ export default function Register() {
       register(userProfile, password)
         .then(() => history.push("/"));
     }
- };
+  };
 
   return (
     <Form onSubmit={registerClick}>
